+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+* [PR #496] Show sent email when recovering password
 * [PR #494] Recover password with cpf
 
 ## [2.1.1] 27/09/2019